+++ board.html
@@ -89,6 +89,6 @@
 
   <!-- Game board end -->
 
-  <script src="./app.js"></script>
+  <script type="module" src="./app.js"></script>
 </body>
 </html>
