+++ index.html
@@ -77,6 +77,6 @@
   Yes, restart
 
    Game board end -->
-  <script src="./app.js"></script>
+  <script type="module" src="./index.js"></script>
 </body>
 </html>
